feat: 검색창 선택 시 검색 화면으로 이동 및 지도로 복귀하는 기능 구현
