+++ README.md
@@ -17,5 +17,5 @@
 ### 2단계 구현 목록
 - [x] 카카오 SDK 적용
 - [x] 앱 실행 시 지도 화면 표시
-- [ ] 검색 창을 선택하면 검색화면으로 이동
-- [ ] 검색 화면에서 뒤로가기를 하면 지도 화면으로 복귀
+- [x] 검색 창을 선택하면 검색화면으로 이동
+- [x] 검색 화면에서 뒤로가기를 하면 지도 화면으로 복귀
