+++ app/src/main/java/campus/tech/kakao/map/MainActivity.kt
@@ -27,7 +27,6 @@ class MainActivity : AppCompatActivity(), DatabaseListener {
     private lateinit var searchResultAdapter: ResultRecyclerAdapter
     private lateinit var searchHistoryAdapter: HistoryRecyclerAdapter
 
-    private lateinit var kakaoMap: MapView
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -39,7 +38,6 @@ class MainActivity : AppCompatActivity(), DatabaseListener {
         searchResultView = findViewById(R.id.search_result)
         message = findViewById(R.id.message)
         clear = findViewById(R.id.clear)
-        kakaoMap = findViewById(R.id.kakao_map)
 
         searchBox.doAfterTextChanged { text ->
             text?.let {
@@ -58,27 +56,6 @@ class MainActivity : AppCompatActivity(), DatabaseListener {
         initSearchResultView()
         initSearchHistoryView()
         observeData()
-        kakaoMap.start(object : MapLifeCycleCallback() {
-            override fun onMapDestroy() {
-            }
-
-            override fun onMapError(p0: Exception?) {
-            }
-
-        }, object : KakaoMapReadyCallback(){
-            override fun onMapReady(p0: KakaoMap) {
-            }
-        })
-    }
-
-    override fun onResume() {
-        super.onResume()
-        kakaoMap.resume()
-    }
-
-    override fun onPause() {
-        super.onPause()
-        kakaoMap.pause()
     }
 
     override fun deleteHistory(historyName: String) {
